# specs/core/openapi.yaml
openapi: 3.0.3
info:
  title: AGRO MIRAI API
  version: v1
  description: |
    Contract skeleton for the AGRO MIRAI HTTP API. This document defines
    the shape every endpoint will honour once Module 11 implements them;
    the frontend (Module 14) may type against it beforehand.

    Conventions: see docs/conventions.md. Enums: see specs/core/enums.md.
    Entity schemas mirror specs/core/schema.yaml.
servers:
  - url: https://api.agromirai.local/v1
    description: Placeholder; real hosts land in Module 11.
security:
  - bearerAuth: []
paths:
...
  /fields/{field_id}/recommendation:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get the most recent CropRecommendation for a Field.
      responses:
        '200':
          description: The recommendation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CropRecommendation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /fields/{field_id}/irrigation:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get the most recent IrrigationAdvice for a Field.
      responses:
        '200':
          description: The advice.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IrrigationAdvice' }
        '404': { $ref: '#/components/responses/NotFound' }
  /fields/{field_id}/disease-risk:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: List DiseaseRiskAlerts for a Field.
      responses:
        '200':
          description: A list of alerts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DiseaseRiskAlert' }
                required: [items]
  /fields/{field_id}/disease-risk/image:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    post:
      summary: |
        Module 21. Submit a leaf photo for CNN-based disease-risk
        classification. Calls the standalone CNN inference service
        (CNN_SERVICE_URL) with a bounded timeout; on any failure to
        reach it (unreachable, timeout, error), falls back to the
        existing rule-based environmental DiseaseRiskModel and still
        returns a valid 200 DiseaseRiskAlert — never a 500. The
        resulting alert's `source` field ("cnn" |
        "environmental_fallback") records which path was actually used.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG/PNG/WEBP leaf photo, max 10MB.
              required: [image]
      responses:
        '200':
          description: The resulting alert (CNN or environmental-fallback path).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiseaseRiskAlert' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /fields/{field_id}/advisories:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: List Advisories for a Field.
      responses:
        '200':
          description: A list of advisories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Advisory' }
                required: [items]
  /feedback:
    post:
      summary: Submit farmer feedback on an Advisory.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeedbackEntryCreate' }
      responses:
        '201':
          description: The stored feedback entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedbackEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --- v2: multi-tenant session auth + read-only Admin (Module 19) ---
  # Session-cookie authenticated, not bearerAuth — see sessionCookie below.
  # decisions/0017-multi-tenant-v2.md. /v1 above is unchanged and stays
  # live for demo/backward-compat; these are additive new paths, not a
  # revision of the /v1 ones. Module 27: farmer auth is Name+Phone+OTP
  # (request-otp/verify-otp), not email+password — see
  # decisions/0023-name-phone-otp-auth.md. The admin-account
  # email+password login stays on /admin/login (server-rendered, not in
  # this API surface) since admin accounts are provisioned out-of-band.
...
  /v2/fields/{field_id}/recommendation:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get the most recent CropRecommendation for a Field owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: The recommendation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CropRecommendation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /v2/fields/{field_id}/irrigation:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get the most recent IrrigationAdvice for a Field owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: The advice.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IrrigationAdvice' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /v2/fields/{field_id}/disease-risk:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: List DiseaseRiskAlerts for a Field owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: A list of alerts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DiseaseRiskAlert' }
                required: [items]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /v2/fields/{field_id}/disease-risk/image:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    post:
      summary: Session-authenticated equivalent of POST /fields/{field_id}/disease-risk/image (Module 21). Same CNN-vs-environmental-fallback contract — see that path's description.
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG/PNG/WEBP leaf photo, max 10MB.
              required: [image]
      responses:
        '200':
          description: The resulting alert (CNN or environmental-fallback path).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiseaseRiskAlert' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /v2/fields/{field_id}/advisories:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      x-note-module-39: >-
        Optional query `generate` (default true). `generate=false` only lists
        the advisories already stored and does not compute a new one.
      summary: List Advisories for a Field owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: A list of advisories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Advisory' }
                required: [items]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /v2/feedback:
    post:
      summary: Submit farmer feedback on an Advisory owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeedbackEntryCreate' }
      responses:
        '201':
          description: The stored feedback entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedbackEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v2/bug-reports:
    post:
      summary: Submit a farmer-facing in-app bug report (not tied to a specific Advisory).
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BugReportCreate' }
      responses:
        '201':
          description: The stored bug report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BugReport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask signed session cookie issued by POST /v2/auth/login (Module 19). Not a bearer token — the cookie is opaque and itsdangerous-signed.
  parameters:
    FieldId:
      name: field_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found or not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    UnprocessableEntity:
      description: Well-formed request, but the field's data isn't sufficient to compute a result (e.g. no weather data yet).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    CropRecommendation:
      type: object
      required: [id, field_id, created_at, recommended_crop, confidence]
      properties:
        id: { type: string, format: uuid }
        field_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        recommended_crop:
          type: string
          enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
        confidence: { type: number, minimum: 0, maximum: 1 }
        alternatives:
          type: array
          items:
            type: string
            enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
        rationale: { type: string }
        season:
          type: string
          enum: [kharif, rabi, zaid]
        out_of_region: { type: boolean }
        regional_alternative:
          type: string
          enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
    IrrigationAdvice:
      type: object
      required: [id, field_id, created_at, recommended_depth_mm, window_start_at, window_end_at, urgency]
      properties:
        id: { type: string, format: uuid }
        field_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        recommended_depth_mm: { type: number, minimum: 0 }
        window_start_at: { type: string, format: date-time }
        window_end_at: { type: string, format: date-time }
        urgency:
          type: string
          enum: [low, moderate, high, severe]
        rationale: { type: string }
    DiseaseRiskAlert:
      type: object
      required: [id, field_id, created_at, disease, risk_level, confidence]
      properties:
        id: { type: string, format: uuid }
        field_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        disease: { type: string }
        risk_level:
          type: string
          enum: [low, moderate, high, severe]
        confidence: { type: number, minimum: 0, maximum: 1 }
        window_start_at: { type: string, format: date-time }
        window_end_at: { type: string, format: date-time }
        recommended_action: { type: string }
        source:
          type: string
          description: Module 21 additive. Which path produced this alert.
          enum: [cnn, environmental, environmental_fallback]
    Advisory:
      type: object
      required: [id, field_id, created_at, language, title, body, severity]
      properties:
        id: { type: string, format: uuid }
        field_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        language:
          type: string
          enum: [en, hi, kn, ta, te, mr, bn, gu]
        title: { type: string }
        body: { type: string }
        body_plain:
          type: string
          description: Plain farmer-language rewrite of `body` (response only).
        severity:
          type: string
          enum: [low, moderate, high, severe]
        source_refs:
          type: array
          items: { type: string, format: uuid }
    FeedbackEntry:
      type: object
      required: [id, farmer_id, advisory_id, created_at, rating, helpful]
      properties:
        id: { type: string, format: uuid }
        farmer_id: { type: string, format: uuid }
        advisory_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        rating: { type: integer, minimum: 1, maximum: 5 }
        helpful: { type: boolean }
        comment: { type: string }
    FeedbackEntryCreate:
      type: object
      required: [advisory_id, rating, helpful]
      properties:
        advisory_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        helpful: { type: boolean }
        comment: { type: string }
    BugReport:
      type: object
      required: [id, farmer_id, created_at]
      properties:
        id: { type: string, format: uuid }
        farmer_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        category:
          type: string
          enum: [crash, wrong_info, unclear_advice, photo_scan_failed, login_failed, other]
        message: { type: string }
        photo_url: { type: string }
        app_version: { type: string }
        platform: { type: string }
    BugReportCreate:
      type: object
      description: At least one of category/message is required (enforced server-side, not by an "at least one" JSON Schema construct here).
      properties:
        category:
          type: string
          enum: [crash, wrong_info, unclear_advice, photo_scan_failed, login_failed, other]
        message: { type: string }
        photo_url: { type: string }
        app_version: { type: string }
        platform: { type: string }